BackEnd/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.23
django-cors-headers==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2